backend/sentiment_analysis/serializer.py
# sentiment_analysis/serializers.py

from rest_framework import serializers
from .models import ProductReview

class ProductReviewSerializer(serializers.ModelSerializer):
    class Meta:
        model = ProductReview
        fields = '__all__'