backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
pymongo==4.5.0
djongo==1.3.6
scikit-learn==1.3.0
numpy==1.25.2
pandas==2.1.0
matplotlib==3.7.2
nltk==3.8.1
textblob==0.15.3
python-dotenv==1.0.0
gunicorn==21.2.0
bson==0.5.10